+++ CHANGELOG.md
@@ -1,6 +1,7 @@
-## 1.5.0 [In progress]
+## 2.0.0 [In progress]
 ### Breaking changes
-1. [#173](https://github.com/influxdata/influxdb-client-go/pull/173) Removed orgs labels API cause [it has been removed from the server API](https://github.com/influxdata/influxdb/pull/19104)
+1. [#173](https://github.com/influxdata/influxdb-client-go/pull/173) Removed deprecated API.
+1. [#174](https://github.com/influxdata/influxdb-client-go/pull/174) Removed orgs labels API cause [it has been removed from the server API](https://github.com/influxdata/influxdb/pull/19104)
 
 ### Features
 1. [#165](https://github.com/influxdata/influxdb-client-go/pull/165) Allow overriding the http.Client for the http service.
@@ -61,6 +62,6 @@
 
 - initial release of new client version
 
-### Apis
+### APIs
 
 - initial release of new client version